Add Dto folder
